openai-chatgpt/requirements.txt
python-dotenv==1.1.1
langchain==1.0.2
langchain-chroma==1.0.0
langchain-community==0.4
langchain-openai==1.0.1
chromadb==1.2.1
langchain-ollama==1.0.0
openai==2.6.0
pypdf==6.1.2